common-prefix-search関数周りを少し整理

+++ analyzer/src/net/reduls/gomoku/dic/Matrix.java
@@ -20,7 +20,7 @@ public final class Matrix {
         Misc.close(in);
     }
 
-    public static short linkCost(int leftId, int rightId) {
+    public static short linkCost(short leftId, short rightId) {
         return matrix[rightId][leftId];
     }
 }
